+++ server/impl/src/test/java/com/walmartlabs/concord/server/process/pipelines/processors/RequestDataMergingProcessorTest.java
@@ -23,9 +23,9 @@ package com.walmartlabs.concord.server.process.pipelines.processors;
 import com.walmartlabs.concord.project.InternalConstants;
 import com.walmartlabs.concord.server.cfg.DefaultVariablesConfiguration;
 import com.walmartlabs.concord.server.org.OrganizationDao;
+import com.walmartlabs.concord.server.org.policy.PolicyEntry;
 import com.walmartlabs.concord.server.org.project.ProjectDao;
 import com.walmartlabs.concord.server.process.Payload;
-import com.walmartlabs.concord.server.process.PolicyReader;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.stubbing.Answer;
@@ -48,16 +48,14 @@ public class RequestDataMergingProcessorTest {
     private RequestDataMergingProcessor p;
     private ProjectDao projectDao;
     private OrganizationDao orgDao;
-    private PolicyReader policyReader;
 
     @Before
     public void init() {
         projectDao = mock(ProjectDao.class);
         orgDao = mock(OrganizationDao.class);
-        policyReader = mock(PolicyReader.class);
 
         DefaultVariablesConfiguration defaultVars = mock(DefaultVariablesConfiguration.class);
-        p = new RequestDataMergingProcessor(projectDao, orgDao, defaultVars, policyReader);
+        p = new RequestDataMergingProcessor(projectDao, orgDao, defaultVars);
     }
 
     @Test
@@ -90,14 +88,14 @@ public class RequestDataMergingProcessorTest {
         // ---
         when(orgDao.getConfiguration(eq(orgId))).thenReturn(orgCfg);
         when(projectDao.getConfiguration(eq(prjId))).thenReturn(prjCfg);
-        when(policyReader.readPolicy(any())).thenReturn(policy);
 
         Payload payload = new Payload(instanceId);
         payload = payload
                 .putHeader(Payload.REQUEST_DATA_MAP, req)
                 .putHeader(Payload.ORGANIZATION_ID, orgId)
                 .putHeader(Payload.PROJECT_ID, prjId)
-                .putHeader(Payload.WORKSPACE_DIR, workDir);
+                .putHeader(Payload.WORKSPACE_DIR, workDir)
+                .putHeader(Payload.POLICY, new PolicyEntry("test", policy));
 
         // ---
         Map<String, Object> expected = new HashMap<>();